@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -10,7 +10,7 @@ package armsynapse
 
 const (
 	moduleName    = "armsynapse"
-	moduleVersion = "v0.2.1"
+	moduleVersion = "v0.3.0"
 )
 
 type AzureADOnlyAuthenticationName string
@@ -24,11 +24,6 @@ func PossibleAzureADOnlyAuthenticationNameValues() []AzureADOnlyAuthenticationNa
 	return []AzureADOnlyAuthenticationName{
 		AzureADOnlyAuthenticationNameDefault,
 	}
-}
-
-// ToPtr returns a *AzureADOnlyAuthenticationName pointing to the current value.
-func (c AzureADOnlyAuthenticationName) ToPtr() *AzureADOnlyAuthenticationName {
-	return &c
 }
 
 // AzureScaleType - Scale type.
@@ -49,11 +44,6 @@ func PossibleAzureScaleTypeValues() []AzureScaleType {
 	}
 }
 
-// ToPtr returns a *AzureScaleType pointing to the current value.
-func (c AzureScaleType) ToPtr() *AzureScaleType {
-	return &c
-}
-
 type BlobAuditingPolicyName string
 
 const (
@@ -65,11 +55,6 @@ func PossibleBlobAuditingPolicyNameValues() []BlobAuditingPolicyName {
 	return []BlobAuditingPolicyName{
 		BlobAuditingPolicyNameDefault,
 	}
-}
-
-// ToPtr returns a *BlobAuditingPolicyName pointing to the current value.
-func (c BlobAuditingPolicyName) ToPtr() *BlobAuditingPolicyName {
-	return &c
 }
 
 // BlobAuditingPolicyState - Specifies the state of the policy. If state is Enabled, storageEndpoint or isAzureMonitorTargetEnabled
@@ -89,11 +74,6 @@ func PossibleBlobAuditingPolicyStateValues() []BlobAuditingPolicyState {
 	}
 }
 
-// ToPtr returns a *BlobAuditingPolicyState pointing to the current value.
-func (c BlobAuditingPolicyState) ToPtr() *BlobAuditingPolicyState {
-	return &c
-}
-
 // BlobStorageEventType - The name of blob storage event type to process.
 type BlobStorageEventType string
 
@@ -110,11 +90,6 @@ func PossibleBlobStorageEventTypeValues() []BlobStorageEventType {
 	}
 }
 
-// ToPtr returns a *BlobStorageEventType pointing to the current value.
-func (c BlobStorageEventType) ToPtr() *BlobStorageEventType {
-	return &c
-}
-
 // ClusterPrincipalRole - Cluster principal role.
 type ClusterPrincipalRole string
 
@@ -129,11 +104,6 @@ func PossibleClusterPrincipalRoleValues() []ClusterPrincipalRole {
 		ClusterPrincipalRoleAllDatabasesAdmin,
 		ClusterPrincipalRoleAllDatabasesViewer,
 	}
-}
-
-// ToPtr returns a *ClusterPrincipalRole pointing to the current value.
-func (c ClusterPrincipalRole) ToPtr() *ClusterPrincipalRole {
-	return &c
 }
 
 // ColumnDataType - The column data type.
@@ -216,11 +186,6 @@ func PossibleColumnDataTypeValues() []ColumnDataType {
 	}
 }
 
-// ToPtr returns a *ColumnDataType pointing to the current value.
-func (c ColumnDataType) ToPtr() *ColumnDataType {
-	return &c
-}
-
 // Compression - The compression type
 type Compression string
 
@@ -235,11 +200,6 @@ func PossibleCompressionValues() []Compression {
 		CompressionGZip,
 		CompressionNone,
 	}
-}
-
-// ToPtr returns a *Compression pointing to the current value.
-func (c Compression) ToPtr() *Compression {
-	return &c
 }
 
 // ConfigurationType - The type of the spark config properties file.
@@ -258,11 +218,6 @@ func PossibleConfigurationTypeValues() []ConfigurationType {
 	}
 }
 
-// ToPtr returns a *ConfigurationType pointing to the current value.
-func (c ConfigurationType) ToPtr() *ConfigurationType {
-	return &c
-}
-
 type ConnectionPolicyName string
 
 const (
@@ -274,11 +229,6 @@ func PossibleConnectionPolicyNameValues() []ConnectionPolicyName {
 	return []ConnectionPolicyName{
 		ConnectionPolicyNameDefault,
 	}
-}
-
-// ToPtr returns a *ConnectionPolicyName pointing to the current value.
-func (c ConnectionPolicyName) ToPtr() *ConnectionPolicyName {
-	return &c
 }
 
 // CreateMode - Specifies the mode of sql pool creation.
@@ -310,11 +260,6 @@ func PossibleCreateModeValues() []CreateMode {
 	}
 }
 
-// ToPtr returns a *CreateMode pointing to the current value.
-func (c CreateMode) ToPtr() *CreateMode {
-	return &c
-}
-
 // CreatedByType - The type of identity that created the resource.
 type CreatedByType string
 
@@ -335,11 +280,6 @@ func PossibleCreatedByTypeValues() []CreatedByType {
 	}
 }
 
-// ToPtr returns a *CreatedByType pointing to the current value.
-func (c CreatedByType) ToPtr() *CreatedByType {
-	return &c
-}
-
 // DataConnectionKind - Kind of the endpoint for the data connection
 type DataConnectionKind string
 
@@ -358,11 +298,6 @@ func PossibleDataConnectionKindValues() []DataConnectionKind {
 	}
 }
 
-// ToPtr returns a *DataConnectionKind pointing to the current value.
-func (c DataConnectionKind) ToPtr() *DataConnectionKind {
-	return &c
-}
-
 // DataFlowComputeType - Compute type of the cluster which will execute data flow job.
 type DataFlowComputeType string
 
@@ -379,11 +314,6 @@ func PossibleDataFlowComputeTypeValues() []DataFlowComputeType {
 		DataFlowComputeTypeGeneral,
 		DataFlowComputeTypeMemoryOptimized,
 	}
-}
-
-// ToPtr returns a *DataFlowComputeType pointing to the current value.
-func (c DataFlowComputeType) ToPtr() *DataFlowComputeType {
-	return &c
 }
 
 // DataMaskingFunction - The masking function that is used for the data masking rule.
@@ -410,11 +340,6 @@ func PossibleDataMaskingFunctionValues() []DataMaskingFunction {
 	}
 }
 
-// ToPtr returns a *DataMaskingFunction pointing to the current value.
-func (c DataMaskingFunction) ToPtr() *DataMaskingFunction {
-	return &c
-}
-
 // DataMaskingRuleState - The rule state. Used to delete a rule. To delete an existing rule, specify the schemaName, tableName,
 // columnName, maskingFunction, and specify ruleState as disabled. However, if the rule doesn't
 // already exist, the rule will be created with ruleState set to enabled, regardless of the provided value of ruleState.
@@ -433,11 +358,6 @@ func PossibleDataMaskingRuleStateValues() []DataMaskingRuleState {
 	}
 }
 
-// ToPtr returns a *DataMaskingRuleState pointing to the current value.
-func (c DataMaskingRuleState) ToPtr() *DataMaskingRuleState {
-	return &c
-}
-
 // DataMaskingState - The state of the data masking policy.
 type DataMaskingState string
 
@@ -454,11 +374,6 @@ func PossibleDataMaskingStateValues() []DataMaskingState {
 	}
 }
 
-// ToPtr returns a *DataMaskingState pointing to the current value.
-func (c DataMaskingState) ToPtr() *DataMaskingState {
-	return &c
-}
-
 type DataWarehouseUserActivityName string
 
 const (
@@ -470,11 +385,6 @@ func PossibleDataWarehouseUserActivityNameValues() []DataWarehouseUserActivityNa
 	return []DataWarehouseUserActivityName{
 		DataWarehouseUserActivityNameCurrent,
 	}
-}
-
-// ToPtr returns a *DataWarehouseUserActivityName pointing to the current value.
-func (c DataWarehouseUserActivityName) ToPtr() *DataWarehouseUserActivityName {
-	return &c
 }
 
 // DatabasePrincipalRole - Database principal role.
@@ -499,11 +409,6 @@ func PossibleDatabasePrincipalRoleValues() []DatabasePrincipalRole {
 		DatabasePrincipalRoleUser,
 		DatabasePrincipalRoleViewer,
 	}
-}
-
-// ToPtr returns a *DatabasePrincipalRole pointing to the current value.
-func (c DatabasePrincipalRole) ToPtr() *DatabasePrincipalRole {
-	return &c
 }
 
 // DayOfWeek - Day of maintenance window.
@@ -532,9 +437,17 @@ func PossibleDayOfWeekValues() []DayOfWeek {
 	}
 }
 
-// ToPtr returns a *DayOfWeek pointing to the current value.
-func (c DayOfWeek) ToPtr() *DayOfWeek {
-	return &c
+type DedicatedSQLMinimalTLSSettingsName string
+
+const (
+	DedicatedSQLMinimalTLSSettingsNameDefault DedicatedSQLMinimalTLSSettingsName = "default"
+)
+
+// PossibleDedicatedSQLMinimalTLSSettingsNameValues returns the possible values for the DedicatedSQLMinimalTLSSettingsName const type.
+func PossibleDedicatedSQLMinimalTLSSettingsNameValues() []DedicatedSQLMinimalTLSSettingsName {
+	return []DedicatedSQLMinimalTLSSettingsName{
+		DedicatedSQLMinimalTLSSettingsNameDefault,
+	}
 }
 
 // DefaultPrincipalsModificationKind - The default principals modification kind
@@ -555,11 +468,6 @@ func PossibleDefaultPrincipalsModificationKindValues() []DefaultPrincipalsModifi
 	}
 }
 
-// ToPtr returns a *DefaultPrincipalsModificationKind pointing to the current value.
-func (c DefaultPrincipalsModificationKind) ToPtr() *DefaultPrincipalsModificationKind {
-	return &c
-}
-
 type EncryptionProtectorName string
 
 const (
@@ -571,29 +479,6 @@ func PossibleEncryptionProtectorNameValues() []EncryptionProtectorName {
 	return []EncryptionProtectorName{
 		EncryptionProtectorNameCurrent,
 	}
-}
-
-// ToPtr returns a *EncryptionProtectorName pointing to the current value.
-func (c EncryptionProtectorName) ToPtr() *EncryptionProtectorName {
-	return &c
-}
-
-type Enum11 string
-
-const (
-	Enum11Default Enum11 = "default"
-)
-
-// PossibleEnum11Values returns the possible values for the Enum11 const type.
-func PossibleEnum11Values() []Enum11 {
-	return []Enum11{
-		Enum11Default,
-	}
-}
-
-// ToPtr returns a *Enum11 pointing to the current value.
-func (c Enum11) ToPtr() *Enum11 {
-	return &c
 }
 
 // EventGridDataFormat - The data format of the message. Optionally the data format can be added to each message.
@@ -640,11 +525,6 @@ func PossibleEventGridDataFormatValues() []EventGridDataFormat {
 	}
 }
 
-// ToPtr returns a *EventGridDataFormat pointing to the current value.
-func (c EventGridDataFormat) ToPtr() *EventGridDataFormat {
-	return &c
-}
-
 // EventHubDataFormat - The data format of the message. Optionally the data format can be added to each message.
 type EventHubDataFormat string
 
@@ -689,11 +569,6 @@ func PossibleEventHubDataFormatValues() []EventHubDataFormat {
 	}
 }
 
-// ToPtr returns a *EventHubDataFormat pointing to the current value.
-func (c EventHubDataFormat) ToPtr() *EventHubDataFormat {
-	return &c
-}
-
 type GeoBackupPolicyName string
 
 const (
@@ -705,11 +580,6 @@ func PossibleGeoBackupPolicyNameValues() []GeoBackupPolicyName {
 	return []GeoBackupPolicyName{
 		GeoBackupPolicyNameDefault,
 	}
-}
-
-// ToPtr returns a *GeoBackupPolicyName pointing to the current value.
-func (c GeoBackupPolicyName) ToPtr() *GeoBackupPolicyName {
-	return &c
 }
 
 // GeoBackupPolicyState - The state of the geo backup policy.
@@ -728,11 +598,6 @@ func PossibleGeoBackupPolicyStateValues() []GeoBackupPolicyState {
 	}
 }
 
-// ToPtr returns a *GeoBackupPolicyState pointing to the current value.
-func (c GeoBackupPolicyState) ToPtr() *GeoBackupPolicyState {
-	return &c
-}
-
 // IntegrationRuntimeAuthKeyName - The name of the authentication key to regenerate.
 type IntegrationRuntimeAuthKeyName string
 
@@ -747,11 +612,6 @@ func PossibleIntegrationRuntimeAuthKeyNameValues() []IntegrationRuntimeAuthKeyNa
 		IntegrationRuntimeAuthKeyNameAuthKey1,
 		IntegrationRuntimeAuthKeyNameAuthKey2,
 	}
-}
-
-// ToPtr returns a *IntegrationRuntimeAuthKeyName pointing to the current value.
-func (c IntegrationRuntimeAuthKeyName) ToPtr() *IntegrationRuntimeAuthKeyName {
-	return &c
 }
 
 // IntegrationRuntimeAutoUpdate - The state of integration runtime auto update.
@@ -770,11 +630,6 @@ func PossibleIntegrationRuntimeAutoUpdateValues() []IntegrationRuntimeAutoUpdate
 	}
 }
 
-// ToPtr returns a *IntegrationRuntimeAutoUpdate pointing to the current value.
-func (c IntegrationRuntimeAutoUpdate) ToPtr() *IntegrationRuntimeAutoUpdate {
-	return &c
-}
-
 // IntegrationRuntimeEdition - The edition for the SSIS Integration Runtime
 type IntegrationRuntimeEdition string
 
@@ -791,11 +646,6 @@ func PossibleIntegrationRuntimeEditionValues() []IntegrationRuntimeEdition {
 	}
 }
 
-// ToPtr returns a *IntegrationRuntimeEdition pointing to the current value.
-func (c IntegrationRuntimeEdition) ToPtr() *IntegrationRuntimeEdition {
-	return &c
-}
-
 // IntegrationRuntimeEntityReferenceType - The type of this referenced entity.
 type IntegrationRuntimeEntityReferenceType string
 
@@ -810,11 +660,6 @@ func PossibleIntegrationRuntimeEntityReferenceTypeValues() []IntegrationRuntimeE
 		IntegrationRuntimeEntityReferenceTypeIntegrationRuntimeReference,
 		IntegrationRuntimeEntityReferenceTypeLinkedServiceReference,
 	}
-}
-
-// ToPtr returns a *IntegrationRuntimeEntityReferenceType pointing to the current value.
-func (c IntegrationRuntimeEntityReferenceType) ToPtr() *IntegrationRuntimeEntityReferenceType {
-	return &c
 }
 
 // IntegrationRuntimeInternalChannelEncryptionMode - It is used to set the encryption mode for node-node communication channel
@@ -836,11 +681,6 @@ func PossibleIntegrationRuntimeInternalChannelEncryptionModeValues() []Integrati
 	}
 }
 
-// ToPtr returns a *IntegrationRuntimeInternalChannelEncryptionMode pointing to the current value.
-func (c IntegrationRuntimeInternalChannelEncryptionMode) ToPtr() *IntegrationRuntimeInternalChannelEncryptionMode {
-	return &c
-}
-
 // IntegrationRuntimeLicenseType - License type for bringing your own license scenario.
 type IntegrationRuntimeLicenseType string
 
@@ -855,11 +695,6 @@ func PossibleIntegrationRuntimeLicenseTypeValues() []IntegrationRuntimeLicenseTy
 		IntegrationRuntimeLicenseTypeBasePrice,
 		IntegrationRuntimeLicenseTypeLicenseIncluded,
 	}
-}
-
-// ToPtr returns a *IntegrationRuntimeLicenseType pointing to the current value.
-func (c IntegrationRuntimeLicenseType) ToPtr() *IntegrationRuntimeLicenseType {
-	return &c
 }
 
 // IntegrationRuntimeSsisCatalogPricingTier - The pricing tier for the catalog database. The valid values could be found in
@@ -881,11 +716,6 @@ func PossibleIntegrationRuntimeSsisCatalogPricingTierValues() []IntegrationRunti
 		IntegrationRuntimeSsisCatalogPricingTierPremiumRS,
 		IntegrationRuntimeSsisCatalogPricingTierStandard,
 	}
-}
-
-// ToPtr returns a *IntegrationRuntimeSsisCatalogPricingTier pointing to the current value.
-func (c IntegrationRuntimeSsisCatalogPricingTier) ToPtr() *IntegrationRuntimeSsisCatalogPricingTier {
-	return &c
 }
 
 // IntegrationRuntimeState - The state of integration runtime.
@@ -920,11 +750,6 @@ func PossibleIntegrationRuntimeStateValues() []IntegrationRuntimeState {
 	}
 }
 
-// ToPtr returns a *IntegrationRuntimeState pointing to the current value.
-func (c IntegrationRuntimeState) ToPtr() *IntegrationRuntimeState {
-	return &c
-}
-
 // IntegrationRuntimeType - The type of integration runtime.
 type IntegrationRuntimeType string
 
@@ -939,11 +764,6 @@ func PossibleIntegrationRuntimeTypeValues() []IntegrationRuntimeType {
 		IntegrationRuntimeTypeManaged,
 		IntegrationRuntimeTypeSelfHosted,
 	}
-}
-
-// ToPtr returns a *IntegrationRuntimeType pointing to the current value.
-func (c IntegrationRuntimeType) ToPtr() *IntegrationRuntimeType {
-	return &c
 }
 
 // IntegrationRuntimeUpdateResult - The result of the last integration runtime node update.
@@ -962,11 +782,6 @@ func PossibleIntegrationRuntimeUpdateResultValues() []IntegrationRuntimeUpdateRe
 		IntegrationRuntimeUpdateResultNone,
 		IntegrationRuntimeUpdateResultSucceed,
 	}
-}
-
-// ToPtr returns a *IntegrationRuntimeUpdateResult pointing to the current value.
-func (c IntegrationRuntimeUpdateResult) ToPtr() *IntegrationRuntimeUpdateResult {
-	return &c
 }
 
 // IotHubDataFormat - The data format of the message. Optionally the data format can be added to each message.
@@ -1013,11 +828,6 @@ func PossibleIotHubDataFormatValues() []IotHubDataFormat {
 	}
 }
 
-// ToPtr returns a *IotHubDataFormat pointing to the current value.
-func (c IotHubDataFormat) ToPtr() *IotHubDataFormat {
-	return &c
-}
-
 // Kind - Kind of the database
 type Kind string
 
@@ -1034,11 +844,6 @@ func PossibleKindValues() []Kind {
 	}
 }
 
-// ToPtr returns a *Kind pointing to the current value.
-func (c Kind) ToPtr() *Kind {
-	return &c
-}
-
 // LanguageExtensionName - Language extension that can run within KQL query.
 type LanguageExtensionName string
 
@@ -1053,11 +858,6 @@ func PossibleLanguageExtensionNameValues() []LanguageExtensionName {
 		LanguageExtensionNamePYTHON,
 		LanguageExtensionNameR,
 	}
-}
-
-// ToPtr returns a *LanguageExtensionName pointing to the current value.
-func (c LanguageExtensionName) ToPtr() *LanguageExtensionName {
-	return &c
 }
 
 // ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityActualState - Actual state
@@ -1082,11 +882,6 @@ func PossibleManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToMa
 	}
 }
 
-// ToPtr returns a *ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityActualState pointing to the current value.
-func (c ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityActualState) ToPtr() *ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityActualState {
-	return &c
-}
-
 // ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityDesiredState - Desired state
 type ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityDesiredState string
 
@@ -1101,11 +896,6 @@ func PossibleManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToMa
 		ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityDesiredStateDisabled,
 		ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityDesiredStateEnabled,
 	}
-}
-
-// ToPtr returns a *ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityDesiredState pointing to the current value.
-func (c ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityDesiredState) ToPtr() *ManagedIdentitySQLControlSettingsModelPropertiesGrantSQLControlToManagedIdentityDesiredState {
-	return &c
 }
 
 // ManagedIntegrationRuntimeNodeStatus - The managed integration runtime node status.
@@ -1126,11 +916,6 @@ func PossibleManagedIntegrationRuntimeNodeStatusValues() []ManagedIntegrationRun
 		ManagedIntegrationRuntimeNodeStatusStarting,
 		ManagedIntegrationRuntimeNodeStatusUnavailable,
 	}
-}
-
-// ToPtr returns a *ManagedIntegrationRuntimeNodeStatus pointing to the current value.
-func (c ManagedIntegrationRuntimeNodeStatus) ToPtr() *ManagedIntegrationRuntimeNodeStatus {
-	return &c
 }
 
 // ManagementOperationState - The operation state.
@@ -1155,11 +940,6 @@ func PossibleManagementOperationStateValues() []ManagementOperationState {
 		ManagementOperationStatePending,
 		ManagementOperationStateSucceeded,
 	}
-}
-
-// ToPtr returns a *ManagementOperationState pointing to the current value.
-func (c ManagementOperationState) ToPtr() *ManagementOperationState {
-	return &c
 }
 
 // NodeSize - The level of compute power that each node in the Big Data pool has.
@@ -1188,11 +968,6 @@ func PossibleNodeSizeValues() []NodeSize {
 	}
 }
 
-// ToPtr returns a *NodeSize pointing to the current value.
-func (c NodeSize) ToPtr() *NodeSize {
-	return &c
-}
-
 // NodeSizeFamily - The kind of nodes that the Big Data pool provides.
 type NodeSizeFamily string
 
@@ -1211,11 +986,6 @@ func PossibleNodeSizeFamilyValues() []NodeSizeFamily {
 		NodeSizeFamilyMemoryOptimized,
 		NodeSizeFamilyNone,
 	}
-}
-
-// ToPtr returns a *NodeSizeFamily pointing to the current value.
-func (c NodeSizeFamily) ToPtr() *NodeSizeFamily {
-	return &c
 }
 
 // OperationStatus - Operation status
@@ -1238,11 +1008,6 @@ func PossibleOperationStatusValues() []OperationStatus {
 	}
 }
 
-// ToPtr returns a *OperationStatus pointing to the current value.
-func (c OperationStatus) ToPtr() *OperationStatus {
-	return &c
-}
-
 // PrincipalType - Principal type.
 type PrincipalType string
 
@@ -1261,11 +1026,6 @@ func PossiblePrincipalTypeValues() []PrincipalType {
 	}
 }
 
-// ToPtr returns a *PrincipalType pointing to the current value.
-func (c PrincipalType) ToPtr() *PrincipalType {
-	return &c
-}
-
 // PrincipalsModificationKind - The principals modification kind of the database
 type PrincipalsModificationKind string
 
@@ -1282,11 +1042,6 @@ func PossiblePrincipalsModificationKindValues() []PrincipalsModificationKind {
 		PrincipalsModificationKindReplace,
 		PrincipalsModificationKindUnion,
 	}
-}
-
-// ToPtr returns a *PrincipalsModificationKind pointing to the current value.
-func (c PrincipalsModificationKind) ToPtr() *PrincipalsModificationKind {
-	return &c
 }
 
 // ProvisioningState - Resource provisioning state
@@ -1311,11 +1066,6 @@ func PossibleProvisioningStateValues() []ProvisioningState {
 	}
 }
 
-// ToPtr returns a *ProvisioningState pointing to the current value.
-func (c ProvisioningState) ToPtr() *ProvisioningState {
-	return &c
-}
-
 // QueryAggregationFunction - The function that is used to aggregate each query's metrics.
 type QueryAggregationFunction string
 
@@ -1334,11 +1084,6 @@ func PossibleQueryAggregationFunctionValues() []QueryAggregationFunction {
 		QueryAggregationFunctionAvg,
 		QueryAggregationFunctionSum,
 	}
-}
-
-// ToPtr returns a *QueryAggregationFunction pointing to the current value.
-func (c QueryAggregationFunction) ToPtr() *QueryAggregationFunction {
-	return &c
 }
 
 // QueryExecutionType - The execution type that is used to filter the query instances that are returned.
@@ -1363,11 +1108,6 @@ func PossibleQueryExecutionTypeValues() []QueryExecutionType {
 	}
 }
 
-// ToPtr returns a *QueryExecutionType pointing to the current value.
-func (c QueryExecutionType) ToPtr() *QueryExecutionType {
-	return &c
-}
-
 // QueryMetricUnit - The unit of measurement
 type QueryMetricUnit string
 
@@ -1384,11 +1124,6 @@ func PossibleQueryMetricUnitValues() []QueryMetricUnit {
 		QueryMetricUnitKB,
 		QueryMetricUnitMicroseconds,
 	}
-}
-
-// ToPtr returns a *QueryMetricUnit pointing to the current value.
-func (c QueryMetricUnit) ToPtr() *QueryMetricUnit {
-	return &c
 }
 
 // QueryObservedMetricType - The type of metric to use for ordering the top metrics.
@@ -1413,11 +1148,6 @@ func PossibleQueryObservedMetricTypeValues() []QueryObservedMetricType {
 	}
 }
 
-// ToPtr returns a *QueryObservedMetricType pointing to the current value.
-func (c QueryObservedMetricType) ToPtr() *QueryObservedMetricType {
-	return &c
-}
-
 // Reason - Message providing the reason why the given name is invalid.
 type Reason string
 
@@ -1434,11 +1164,6 @@ func PossibleReasonValues() []Reason {
 	}
 }
 
-// ToPtr returns a *Reason pointing to the current value.
-func (c Reason) ToPtr() *Reason {
-	return &c
-}
-
 type RecommendedSensitivityLabelUpdateKind string
 
 const (
@@ -1452,11 +1177,6 @@ func PossibleRecommendedSensitivityLabelUpdateKindValues() []RecommendedSensitiv
 		RecommendedSensitivityLabelUpdateKindEnable,
 		RecommendedSensitivityLabelUpdateKindDisable,
 	}
-}
-
-// ToPtr returns a *RecommendedSensitivityLabelUpdateKind pointing to the current value.
-func (c RecommendedSensitivityLabelUpdateKind) ToPtr() *RecommendedSensitivityLabelUpdateKind {
-	return &c
 }
 
 // ReplicationRole - The role of the Sql pool in the replication link.
@@ -1481,11 +1201,6 @@ func PossibleReplicationRoleValues() []ReplicationRole {
 	}
 }
 
-// ToPtr returns a *ReplicationRole pointing to the current value.
-func (c ReplicationRole) ToPtr() *ReplicationRole {
-	return &c
-}
-
 // ReplicationState - The replication state for the replication link.
 type ReplicationState string
 
@@ -1506,11 +1221,6 @@ func PossibleReplicationStateValues() []ReplicationState {
 	}
 }
 
-// ToPtr returns a *ReplicationState pointing to the current value.
-func (c ReplicationState) ToPtr() *ReplicationState {
-	return &c
-}
-
 // ResourceIdentityType - The type of managed identity for the workspace
 type ResourceIdentityType string
 
@@ -1527,11 +1237,6 @@ func PossibleResourceIdentityTypeValues() []ResourceIdentityType {
 		ResourceIdentityTypeSystemAssigned,
 		ResourceIdentityTypeSystemAssignedUserAssigned,
 	}
-}
-
-// ToPtr returns a *ResourceIdentityType pointing to the current value.
-func (c ResourceIdentityType) ToPtr() *ResourceIdentityType {
-	return &c
 }
 
 // ResourceProvisioningState - The provisioned state of the resource.
@@ -1560,11 +1265,6 @@ func PossibleResourceProvisioningStateValues() []ResourceProvisioningState {
 	}
 }
 
-// ToPtr returns a *ResourceProvisioningState pointing to the current value.
-func (c ResourceProvisioningState) ToPtr() *ResourceProvisioningState {
-	return &c
-}
-
 // RestorePointType - The type of restore point
 type RestorePointType string
 
@@ -1581,11 +1281,6 @@ func PossibleRestorePointTypeValues() []RestorePointType {
 	}
 }
 
-// ToPtr returns a *RestorePointType pointing to the current value.
-func (c RestorePointType) ToPtr() *RestorePointType {
-	return &c
-}
-
 // SKUName - SKU name.
 type SKUName string
 
@@ -1600,11 +1295,6 @@ func PossibleSKUNameValues() []SKUName {
 		SKUNameComputeOptimized,
 		SKUNameStorageOptimized,
 	}
-}
-
-// ToPtr returns a *SKUName pointing to the current value.
-func (c SKUName) ToPtr() *SKUName {
-	return &c
 }
 
 // SKUSize - SKU size.
@@ -1627,11 +1317,6 @@ func PossibleSKUSizeValues() []SKUSize {
 	}
 }
 
-// ToPtr returns a *SKUSize pointing to the current value.
-func (c SKUSize) ToPtr() *SKUSize {
-	return &c
-}
-
 type SecurityAlertPolicyName string
 
 const (
@@ -1645,11 +1330,6 @@ func PossibleSecurityAlertPolicyNameValues() []SecurityAlertPolicyName {
 	}
 }
 
-// ToPtr returns a *SecurityAlertPolicyName pointing to the current value.
-func (c SecurityAlertPolicyName) ToPtr() *SecurityAlertPolicyName {
-	return &c
-}
-
 type SecurityAlertPolicyNameAutoGenerated string
 
 const (
@@ -1661,11 +1341,6 @@ func PossibleSecurityAlertPolicyNameAutoGeneratedValues() []SecurityAlertPolicyN
 	return []SecurityAlertPolicyNameAutoGenerated{
 		SecurityAlertPolicyNameAutoGeneratedDefault,
 	}
-}
-
-// ToPtr returns a *SecurityAlertPolicyNameAutoGenerated pointing to the current value.
-func (c SecurityAlertPolicyNameAutoGenerated) ToPtr() *SecurityAlertPolicyNameAutoGenerated {
-	return &c
 }
 
 // SecurityAlertPolicyState - Specifies the state of the policy, whether it is enabled or disabled or a policy has not been
@@ -1685,11 +1360,6 @@ func PossibleSecurityAlertPolicyStateValues() []SecurityAlertPolicyState {
 		SecurityAlertPolicyStateEnabled,
 		SecurityAlertPolicyStateDisabled,
 	}
-}
-
-// ToPtr returns a *SecurityAlertPolicyState pointing to the current value.
-func (c SecurityAlertPolicyState) ToPtr() *SecurityAlertPolicyState {
-	return &c
 }
 
 // SelfHostedIntegrationRuntimeNodeStatus - Status of the integration runtime node.
@@ -1718,11 +1388,6 @@ func PossibleSelfHostedIntegrationRuntimeNodeStatusValues() []SelfHostedIntegrat
 	}
 }
 
-// ToPtr returns a *SelfHostedIntegrationRuntimeNodeStatus pointing to the current value.
-func (c SelfHostedIntegrationRuntimeNodeStatus) ToPtr() *SelfHostedIntegrationRuntimeNodeStatus {
-	return &c
-}
-
 type SensitivityLabelRank string
 
 const (
@@ -1744,11 +1409,6 @@ func PossibleSensitivityLabelRankValues() []SensitivityLabelRank {
 	}
 }
 
-// ToPtr returns a *SensitivityLabelRank pointing to the current value.
-func (c SensitivityLabelRank) ToPtr() *SensitivityLabelRank {
-	return &c
-}
-
 type SensitivityLabelSource string
 
 const (
@@ -1762,11 +1422,6 @@ func PossibleSensitivityLabelSourceValues() []SensitivityLabelSource {
 		SensitivityLabelSourceCurrent,
 		SensitivityLabelSourceRecommended,
 	}
-}
-
-// ToPtr returns a *SensitivityLabelSource pointing to the current value.
-func (c SensitivityLabelSource) ToPtr() *SensitivityLabelSource {
-	return &c
 }
 
 type SensitivityLabelUpdateKind string
@@ -1784,11 +1439,6 @@ func PossibleSensitivityLabelUpdateKindValues() []SensitivityLabelUpdateKind {
 	}
 }
 
-// ToPtr returns a *SensitivityLabelUpdateKind pointing to the current value.
-func (c SensitivityLabelUpdateKind) ToPtr() *SensitivityLabelUpdateKind {
-	return &c
-}
-
 // ServerKeyType - The encryption protector type like 'ServiceManaged', 'AzureKeyVault'.
 type ServerKeyType string
 
@@ -1803,11 +1453,6 @@ func PossibleServerKeyTypeValues() []ServerKeyType {
 		ServerKeyTypeAzureKeyVault,
 		ServerKeyTypeServiceManaged,
 	}
-}
-
-// ToPtr returns a *ServerKeyType pointing to the current value.
-func (c ServerKeyType) ToPtr() *ServerKeyType {
-	return &c
 }
 
 // SsisObjectMetadataType - The type of SSIS object metadata.
@@ -1828,11 +1473,6 @@ func PossibleSsisObjectMetadataTypeValues() []SsisObjectMetadataType {
 		SsisObjectMetadataTypePackage,
 		SsisObjectMetadataTypeProject,
 	}
-}
-
-// ToPtr returns a *SsisObjectMetadataType pointing to the current value.
-func (c SsisObjectMetadataType) ToPtr() *SsisObjectMetadataType {
-	return &c
 }
 
 // State - The state of the resource.
@@ -1865,11 +1505,6 @@ func PossibleStateValues() []State {
 	}
 }
 
-// ToPtr returns a *State pointing to the current value.
-func (c State) ToPtr() *State {
-	return &c
-}
-
 // StateValue - property configuration state
 type StateValue string
 
@@ -1888,11 +1523,6 @@ func PossibleStateValueValues() []StateValue {
 	}
 }
 
-// ToPtr returns a *StateValue pointing to the current value.
-func (c StateValue) ToPtr() *StateValue {
-	return &c
-}
-
 // StorageAccountType - The storage account type used to store backups for this sql pool.
 type StorageAccountType string
 
@@ -1909,11 +1539,6 @@ func PossibleStorageAccountTypeValues() []StorageAccountType {
 	}
 }
 
-// ToPtr returns a *StorageAccountType pointing to the current value.
-func (c StorageAccountType) ToPtr() *StorageAccountType {
-	return &c
-}
-
 type TransparentDataEncryptionName string
 
 const (
@@ -1925,11 +1550,6 @@ func PossibleTransparentDataEncryptionNameValues() []TransparentDataEncryptionNa
 	return []TransparentDataEncryptionName{
 		TransparentDataEncryptionNameCurrent,
 	}
-}
-
-// ToPtr returns a *TransparentDataEncryptionName pointing to the current value.
-func (c TransparentDataEncryptionName) ToPtr() *TransparentDataEncryptionName {
-	return &c
 }
 
 // TransparentDataEncryptionStatus - The status of the database transparent data encryption.
@@ -1948,11 +1568,6 @@ func PossibleTransparentDataEncryptionStatusValues() []TransparentDataEncryption
 	}
 }
 
-// ToPtr returns a *TransparentDataEncryptionStatus pointing to the current value.
-func (c TransparentDataEncryptionStatus) ToPtr() *TransparentDataEncryptionStatus {
-	return &c
-}
-
 // Type - The type of resource, for instance Microsoft.Synapse/workspaces/kustoPools/databases.
 type Type string
 
@@ -1969,11 +1584,6 @@ func PossibleTypeValues() []Type {
 	}
 }
 
-// ToPtr returns a *Type pointing to the current value.
-func (c Type) ToPtr() *Type {
-	return &c
-}
-
 type VulnerabilityAssessmentName string
 
 const (
@@ -1985,11 +1595,6 @@ func PossibleVulnerabilityAssessmentNameValues() []VulnerabilityAssessmentName {
 	return []VulnerabilityAssessmentName{
 		VulnerabilityAssessmentNameDefault,
 	}
-}
-
-// ToPtr returns a *VulnerabilityAssessmentName pointing to the current value.
-func (c VulnerabilityAssessmentName) ToPtr() *VulnerabilityAssessmentName {
-	return &c
 }
 
 type VulnerabilityAssessmentPolicyBaselineName string
@@ -2005,11 +1610,6 @@ func PossibleVulnerabilityAssessmentPolicyBaselineNameValues() []VulnerabilityAs
 		VulnerabilityAssessmentPolicyBaselineNameMaster,
 		VulnerabilityAssessmentPolicyBaselineNameDefault,
 	}
-}
-
-// ToPtr returns a *VulnerabilityAssessmentPolicyBaselineName pointing to the current value.
-func (c VulnerabilityAssessmentPolicyBaselineName) ToPtr() *VulnerabilityAssessmentPolicyBaselineName {
-	return &c
 }
 
 // VulnerabilityAssessmentScanState - The scan status.
@@ -2032,11 +1632,6 @@ func PossibleVulnerabilityAssessmentScanStateValues() []VulnerabilityAssessmentS
 	}
 }
 
-// ToPtr returns a *VulnerabilityAssessmentScanState pointing to the current value.
-func (c VulnerabilityAssessmentScanState) ToPtr() *VulnerabilityAssessmentScanState {
-	return &c
-}
-
 // VulnerabilityAssessmentScanTriggerType - The scan trigger type.
 type VulnerabilityAssessmentScanTriggerType string
 
@@ -2053,11 +1648,6 @@ func PossibleVulnerabilityAssessmentScanTriggerTypeValues() []VulnerabilityAsses
 	}
 }
 
-// ToPtr returns a *VulnerabilityAssessmentScanTriggerType pointing to the current value.
-func (c VulnerabilityAssessmentScanTriggerType) ToPtr() *VulnerabilityAssessmentScanTriggerType {
-	return &c
-}
-
 // WorkspacePublicNetworkAccess - Enable or Disable public network access to workspace
 type WorkspacePublicNetworkAccess string
 
@@ -2072,9 +1662,4 @@ func PossibleWorkspacePublicNetworkAccessValues() []WorkspacePublicNetworkAccess
 		WorkspacePublicNetworkAccessDisabled,
 		WorkspacePublicNetworkAccessEnabled,
 	}
-}
-
-// ToPtr returns a *WorkspacePublicNetworkAccess pointing to the current value.
-func (c WorkspacePublicNetworkAccess) ToPtr() *WorkspacePublicNetworkAccess {
-	return &c
 }
