@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type DataMaskingRulesClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewDataMaskingRulesClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *DataMaskingRulesClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewDataMaskingRulesClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*DataMaskingRulesClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &DataMaskingRulesClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CreateOrUpdate - Creates or updates a Sql pool data masking rule.
@@ -110,7 +115,7 @@ func (client *DataMaskingRulesClient) createOrUpdateCreateRequest(ctx context.Co
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
 func (client *DataMaskingRulesClient) createOrUpdateHandleResponse(resp *http.Response) (DataMaskingRulesClientCreateOrUpdateResponse, error) {
-	result := DataMaskingRulesClientCreateOrUpdateResponse{RawResponse: resp}
+	result := DataMaskingRulesClientCreateOrUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.DataMaskingRule); err != nil {
 		return DataMaskingRulesClientCreateOrUpdateResponse{}, err
 	}
@@ -176,7 +181,7 @@ func (client *DataMaskingRulesClient) getCreateRequest(ctx context.Context, reso
 
 // getHandleResponse handles the Get response.
 func (client *DataMaskingRulesClient) getHandleResponse(resp *http.Response) (DataMaskingRulesClientGetResponse, error) {
-	result := DataMaskingRulesClientGetResponse{RawResponse: resp}
+	result := DataMaskingRulesClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.DataMaskingRule); err != nil {
 		return DataMaskingRulesClientGetResponse{}, err
 	}
@@ -190,19 +195,26 @@ func (client *DataMaskingRulesClient) getHandleResponse(resp *http.Response) (Da
 // sqlPoolName - SQL pool name
 // options - DataMaskingRulesClientListBySQLPoolOptions contains the optional parameters for the DataMaskingRulesClient.ListBySQLPool
 // method.
-func (client *DataMaskingRulesClient) ListBySQLPool(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, options *DataMaskingRulesClientListBySQLPoolOptions) (DataMaskingRulesClientListBySQLPoolResponse, error) {
-	req, err := client.listBySQLPoolCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
-	if err != nil {
-		return DataMaskingRulesClientListBySQLPoolResponse{}, err
-	}
-	resp, err := client.pl.Do(req)
-	if err != nil {
-		return DataMaskingRulesClientListBySQLPoolResponse{}, err
-	}
-	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		return DataMaskingRulesClientListBySQLPoolResponse{}, runtime.NewResponseError(resp)
-	}
-	return client.listBySQLPoolHandleResponse(resp)
+func (client *DataMaskingRulesClient) ListBySQLPool(resourceGroupName string, workspaceName string, sqlPoolName string, options *DataMaskingRulesClientListBySQLPoolOptions) *runtime.Pager[DataMaskingRulesClientListBySQLPoolResponse] {
+	return runtime.NewPager(runtime.PageProcessor[DataMaskingRulesClientListBySQLPoolResponse]{
+		More: func(page DataMaskingRulesClientListBySQLPoolResponse) bool {
+			return false
+		},
+		Fetcher: func(ctx context.Context, page *DataMaskingRulesClientListBySQLPoolResponse) (DataMaskingRulesClientListBySQLPoolResponse, error) {
+			req, err := client.listBySQLPoolCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
+			if err != nil {
+				return DataMaskingRulesClientListBySQLPoolResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return DataMaskingRulesClientListBySQLPoolResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return DataMaskingRulesClientListBySQLPoolResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listBySQLPoolHandleResponse(resp)
+		},
+	})
 }
 
 // listBySQLPoolCreateRequest creates the ListBySQLPool request.
@@ -238,7 +250,7 @@ func (client *DataMaskingRulesClient) listBySQLPoolCreateRequest(ctx context.Con
 
 // listBySQLPoolHandleResponse handles the ListBySQLPool response.
 func (client *DataMaskingRulesClient) listBySQLPoolHandleResponse(resp *http.Response) (DataMaskingRulesClientListBySQLPoolResponse, error) {
-	result := DataMaskingRulesClientListBySQLPoolResponse{RawResponse: resp}
+	result := DataMaskingRulesClientListBySQLPoolResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.DataMaskingRuleListResult); err != nil {
 		return DataMaskingRulesClientListBySQLPoolResponse{}, err
 	}
