@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type KeysClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewKeysClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *KeysClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewKeysClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*KeysClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &KeysClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CreateOrUpdate - Creates or updates a workspace key
@@ -103,7 +108,7 @@ func (client *KeysClient) createOrUpdateCreateRequest(ctx context.Context, resou
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
 func (client *KeysClient) createOrUpdateHandleResponse(resp *http.Response) (KeysClientCreateOrUpdateResponse, error) {
-	result := KeysClientCreateOrUpdateResponse{RawResponse: resp}
+	result := KeysClientCreateOrUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Key); err != nil {
 		return KeysClientCreateOrUpdateResponse{}, err
 	}
@@ -163,7 +168,7 @@ func (client *KeysClient) deleteCreateRequest(ctx context.Context, resourceGroup
 
 // deleteHandleResponse handles the Delete response.
 func (client *KeysClient) deleteHandleResponse(resp *http.Response) (KeysClientDeleteResponse, error) {
-	result := KeysClientDeleteResponse{RawResponse: resp}
+	result := KeysClientDeleteResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Key); err != nil {
 		return KeysClientDeleteResponse{}, err
 	}
@@ -223,7 +228,7 @@ func (client *KeysClient) getCreateRequest(ctx context.Context, resourceGroupNam
 
 // getHandleResponse handles the Get response.
 func (client *KeysClient) getHandleResponse(resp *http.Response) (KeysClientGetResponse, error) {
-	result := KeysClientGetResponse{RawResponse: resp}
+	result := KeysClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Key); err != nil {
 		return KeysClientGetResponse{}, err
 	}
@@ -235,16 +240,32 @@ func (client *KeysClient) getHandleResponse(resp *http.Response) (KeysClientGetR
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // workspaceName - The name of the workspace.
 // options - KeysClientListByWorkspaceOptions contains the optional parameters for the KeysClient.ListByWorkspace method.
-func (client *KeysClient) ListByWorkspace(resourceGroupName string, workspaceName string, options *KeysClientListByWorkspaceOptions) *KeysClientListByWorkspacePager {
-	return &KeysClientListByWorkspacePager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listByWorkspaceCreateRequest(ctx, resourceGroupName, workspaceName, options)
+func (client *KeysClient) ListByWorkspace(resourceGroupName string, workspaceName string, options *KeysClientListByWorkspaceOptions) *runtime.Pager[KeysClientListByWorkspaceResponse] {
+	return runtime.NewPager(runtime.PageProcessor[KeysClientListByWorkspaceResponse]{
+		More: func(page KeysClientListByWorkspaceResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp KeysClientListByWorkspaceResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.KeyInfoListResult.NextLink)
+		Fetcher: func(ctx context.Context, page *KeysClientListByWorkspaceResponse) (KeysClientListByWorkspaceResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listByWorkspaceCreateRequest(ctx, resourceGroupName, workspaceName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return KeysClientListByWorkspaceResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return KeysClientListByWorkspaceResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return KeysClientListByWorkspaceResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listByWorkspaceHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listByWorkspaceCreateRequest creates the ListByWorkspace request.
@@ -275,7 +296,7 @@ func (client *KeysClient) listByWorkspaceCreateRequest(ctx context.Context, reso
 
 // listByWorkspaceHandleResponse handles the ListByWorkspace response.
 func (client *KeysClient) listByWorkspaceHandleResponse(resp *http.Response) (KeysClientListByWorkspaceResponse, error) {
-	result := KeysClientListByWorkspaceResponse{RawResponse: resp}
+	result := KeysClientListByWorkspaceResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.KeyInfoListResult); err != nil {
 		return KeysClientListByWorkspaceResponse{}, err
 	}
