@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type OperationsClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewOperationsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *OperationsClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewOperationsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*OperationsClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &OperationsClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CheckNameAvailability - Check whether a workspace name is available
@@ -89,7 +94,7 @@ func (client *OperationsClient) checkNameAvailabilityCreateRequest(ctx context.C
 
 // checkNameAvailabilityHandleResponse handles the CheckNameAvailability response.
 func (client *OperationsClient) checkNameAvailabilityHandleResponse(resp *http.Response) (OperationsClientCheckNameAvailabilityResponse, error) {
-	result := OperationsClientCheckNameAvailabilityResponse{RawResponse: resp}
+	result := OperationsClientCheckNameAvailabilityResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.CheckNameAvailabilityResponse); err != nil {
 		return OperationsClientCheckNameAvailabilityResponse{}, err
 	}
@@ -150,7 +155,7 @@ func (client *OperationsClient) getAzureAsyncHeaderResultCreateRequest(ctx conte
 
 // getAzureAsyncHeaderResultHandleResponse handles the GetAzureAsyncHeaderResult response.
 func (client *OperationsClient) getAzureAsyncHeaderResultHandleResponse(resp *http.Response) (OperationsClientGetAzureAsyncHeaderResultResponse, error) {
-	result := OperationsClientGetAzureAsyncHeaderResultResponse{RawResponse: resp}
+	result := OperationsClientGetAzureAsyncHeaderResultResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.OperationResource); err != nil {
 		return OperationsClientGetAzureAsyncHeaderResultResponse{}, err
 	}
@@ -176,7 +181,7 @@ func (client *OperationsClient) GetLocationHeaderResult(ctx context.Context, res
 	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusCreated, http.StatusAccepted, http.StatusNoContent) {
 		return OperationsClientGetLocationHeaderResultResponse{}, runtime.NewResponseError(resp)
 	}
-	return OperationsClientGetLocationHeaderResultResponse{RawResponse: resp}, nil
+	return OperationsClientGetLocationHeaderResultResponse{}, nil
 }
 
 // getLocationHeaderResultCreateRequest creates the GetLocationHeaderResult request.
@@ -240,7 +245,7 @@ func (client *OperationsClient) listCreateRequest(ctx context.Context, options *
 
 // listHandleResponse handles the List response.
 func (client *OperationsClient) listHandleResponse(resp *http.Response) (OperationsClientListResponse, error) {
-	result := OperationsClientListResponse{RawResponse: resp}
+	result := OperationsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.AvailableRpOperationArray); err != nil {
 		return OperationsClientListResponse{}, err
 	}
