@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type SQLPoolMaintenanceWindowsClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLPoolMaintenanceWindowsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SQLPoolMaintenanceWindowsClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewSQLPoolMaintenanceWindowsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLPoolMaintenanceWindowsClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLPoolMaintenanceWindowsClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CreateOrUpdate - Creates or updates a Sql pool's maintenance windows settings.
@@ -70,7 +75,7 @@ func (client *SQLPoolMaintenanceWindowsClient) CreateOrUpdate(ctx context.Contex
 	if !runtime.HasStatusCode(resp, http.StatusOK) {
 		return SQLPoolMaintenanceWindowsClientCreateOrUpdateResponse{}, runtime.NewResponseError(resp)
 	}
-	return SQLPoolMaintenanceWindowsClientCreateOrUpdateResponse{RawResponse: resp}, nil
+	return SQLPoolMaintenanceWindowsClientCreateOrUpdateResponse{}, nil
 }
 
 // createOrUpdateCreateRequest creates the CreateOrUpdate request.
@@ -159,7 +164,7 @@ func (client *SQLPoolMaintenanceWindowsClient) getCreateRequest(ctx context.Cont
 
 // getHandleResponse handles the Get response.
 func (client *SQLPoolMaintenanceWindowsClient) getHandleResponse(resp *http.Response) (SQLPoolMaintenanceWindowsClientGetResponse, error) {
-	result := SQLPoolMaintenanceWindowsClientGetResponse{RawResponse: resp}
+	result := SQLPoolMaintenanceWindowsClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.MaintenanceWindows); err != nil {
 		return SQLPoolMaintenanceWindowsClientGetResponse{}, err
 	}
