@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type SQLPoolsClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLPoolsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SQLPoolsClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewSQLPoolsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLPoolsClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLPoolsClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // BeginCreate - Create a SQL pool
@@ -56,22 +61,18 @@ func NewSQLPoolsClient(subscriptionID string, credential azcore.TokenCredential,
 // sqlPoolName - SQL pool name
 // sqlPoolInfo - The SQL pool to create
 // options - SQLPoolsClientBeginCreateOptions contains the optional parameters for the SQLPoolsClient.BeginCreate method.
-func (client *SQLPoolsClient) BeginCreate(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, sqlPoolInfo SQLPool, options *SQLPoolsClientBeginCreateOptions) (SQLPoolsClientCreatePollerResponse, error) {
-	resp, err := client.create(ctx, resourceGroupName, workspaceName, sqlPoolName, sqlPoolInfo, options)
-	if err != nil {
-		return SQLPoolsClientCreatePollerResponse{}, err
+func (client *SQLPoolsClient) BeginCreate(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, sqlPoolInfo SQLPool, options *SQLPoolsClientBeginCreateOptions) (*armruntime.Poller[SQLPoolsClientCreateResponse], error) {
+	if options == nil || options.ResumeToken == "" {
+		resp, err := client.create(ctx, resourceGroupName, workspaceName, sqlPoolName, sqlPoolInfo, options)
+		if err != nil {
+			return nil, err
+		}
+		return armruntime.NewPoller(resp, client.pl, &armruntime.NewPollerOptions[SQLPoolsClientCreateResponse]{
+			FinalStateVia: armruntime.FinalStateViaLocation,
+		})
+	} else {
+		return armruntime.NewPollerFromResumeToken[SQLPoolsClientCreateResponse](options.ResumeToken, client.pl, nil)
 	}
-	result := SQLPoolsClientCreatePollerResponse{
-		RawResponse: resp,
-	}
-	pt, err := armruntime.NewPoller("SQLPoolsClient.Create", "location", resp, client.pl)
-	if err != nil {
-		return SQLPoolsClientCreatePollerResponse{}, err
-	}
-	result.Poller = &SQLPoolsClientCreatePoller{
-		pt: pt,
-	}
-	return result, nil
 }
 
 // Create - Create a SQL pool
@@ -127,22 +128,18 @@ func (client *SQLPoolsClient) createCreateRequest(ctx context.Context, resourceG
 // workspaceName - The name of the workspace.
 // sqlPoolName - SQL pool name
 // options - SQLPoolsClientBeginDeleteOptions contains the optional parameters for the SQLPoolsClient.BeginDelete method.
-func (client *SQLPoolsClient) BeginDelete(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolsClientBeginDeleteOptions) (SQLPoolsClientDeletePollerResponse, error) {
-	resp, err := client.deleteOperation(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
-	if err != nil {
-		return SQLPoolsClientDeletePollerResponse{}, err
+func (client *SQLPoolsClient) BeginDelete(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolsClientBeginDeleteOptions) (*armruntime.Poller[SQLPoolsClientDeleteResponse], error) {
+	if options == nil || options.ResumeToken == "" {
+		resp, err := client.deleteOperation(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
+		if err != nil {
+			return nil, err
+		}
+		return armruntime.NewPoller(resp, client.pl, &armruntime.NewPollerOptions[SQLPoolsClientDeleteResponse]{
+			FinalStateVia: armruntime.FinalStateViaLocation,
+		})
+	} else {
+		return armruntime.NewPollerFromResumeToken[SQLPoolsClientDeleteResponse](options.ResumeToken, client.pl, nil)
 	}
-	result := SQLPoolsClientDeletePollerResponse{
-		RawResponse: resp,
-	}
-	pt, err := armruntime.NewPoller("SQLPoolsClient.Delete", "location", resp, client.pl)
-	if err != nil {
-		return SQLPoolsClientDeletePollerResponse{}, err
-	}
-	result.Poller = &SQLPoolsClientDeletePoller{
-		pt: pt,
-	}
-	return result, nil
 }
 
 // Delete - Delete a SQL pool
@@ -245,7 +242,7 @@ func (client *SQLPoolsClient) getCreateRequest(ctx context.Context, resourceGrou
 
 // getHandleResponse handles the Get response.
 func (client *SQLPoolsClient) getHandleResponse(resp *http.Response) (SQLPoolsClientGetResponse, error) {
-	result := SQLPoolsClientGetResponse{RawResponse: resp}
+	result := SQLPoolsClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPool); err != nil {
 		return SQLPoolsClientGetResponse{}, err
 	}
@@ -258,16 +255,32 @@ func (client *SQLPoolsClient) getHandleResponse(resp *http.Response) (SQLPoolsCl
 // workspaceName - The name of the workspace.
 // options - SQLPoolsClientListByWorkspaceOptions contains the optional parameters for the SQLPoolsClient.ListByWorkspace
 // method.
-func (client *SQLPoolsClient) ListByWorkspace(resourceGroupName string, workspaceName string, options *SQLPoolsClientListByWorkspaceOptions) *SQLPoolsClientListByWorkspacePager {
-	return &SQLPoolsClientListByWorkspacePager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listByWorkspaceCreateRequest(ctx, resourceGroupName, workspaceName, options)
+func (client *SQLPoolsClient) ListByWorkspace(resourceGroupName string, workspaceName string, options *SQLPoolsClientListByWorkspaceOptions) *runtime.Pager[SQLPoolsClientListByWorkspaceResponse] {
+	return runtime.NewPager(runtime.PageProcessor[SQLPoolsClientListByWorkspaceResponse]{
+		More: func(page SQLPoolsClientListByWorkspaceResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp SQLPoolsClientListByWorkspaceResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.SQLPoolInfoListResult.NextLink)
+		Fetcher: func(ctx context.Context, page *SQLPoolsClientListByWorkspaceResponse) (SQLPoolsClientListByWorkspaceResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listByWorkspaceCreateRequest(ctx, resourceGroupName, workspaceName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return SQLPoolsClientListByWorkspaceResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return SQLPoolsClientListByWorkspaceResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return SQLPoolsClientListByWorkspaceResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listByWorkspaceHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listByWorkspaceCreateRequest creates the ListByWorkspace request.
@@ -298,7 +311,7 @@ func (client *SQLPoolsClient) listByWorkspaceCreateRequest(ctx context.Context, 
 
 // listByWorkspaceHandleResponse handles the ListByWorkspace response.
 func (client *SQLPoolsClient) listByWorkspaceHandleResponse(resp *http.Response) (SQLPoolsClientListByWorkspaceResponse, error) {
-	result := SQLPoolsClientListByWorkspaceResponse{RawResponse: resp}
+	result := SQLPoolsClientListByWorkspaceResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPoolInfoListResult); err != nil {
 		return SQLPoolsClientListByWorkspaceResponse{}, err
 	}
@@ -311,22 +324,18 @@ func (client *SQLPoolsClient) listByWorkspaceHandleResponse(resp *http.Response)
 // workspaceName - The name of the workspace.
 // sqlPoolName - SQL pool name
 // options - SQLPoolsClientBeginPauseOptions contains the optional parameters for the SQLPoolsClient.BeginPause method.
-func (client *SQLPoolsClient) BeginPause(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolsClientBeginPauseOptions) (SQLPoolsClientPausePollerResponse, error) {
-	resp, err := client.pause(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
-	if err != nil {
-		return SQLPoolsClientPausePollerResponse{}, err
+func (client *SQLPoolsClient) BeginPause(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolsClientBeginPauseOptions) (*armruntime.Poller[SQLPoolsClientPauseResponse], error) {
+	if options == nil || options.ResumeToken == "" {
+		resp, err := client.pause(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
+		if err != nil {
+			return nil, err
+		}
+		return armruntime.NewPoller(resp, client.pl, &armruntime.NewPollerOptions[SQLPoolsClientPauseResponse]{
+			FinalStateVia: armruntime.FinalStateViaLocation,
+		})
+	} else {
+		return armruntime.NewPollerFromResumeToken[SQLPoolsClientPauseResponse](options.ResumeToken, client.pl, nil)
 	}
-	result := SQLPoolsClientPausePollerResponse{
-		RawResponse: resp,
-	}
-	pt, err := armruntime.NewPoller("SQLPoolsClient.Pause", "location", resp, client.pl)
-	if err != nil {
-		return SQLPoolsClientPausePollerResponse{}, err
-	}
-	result.Poller = &SQLPoolsClientPausePoller{
-		pt: pt,
-	}
-	return result, nil
 }
 
 // Pause - Pause a SQL pool
@@ -395,7 +404,7 @@ func (client *SQLPoolsClient) Rename(ctx context.Context, resourceGroupName stri
 	if !runtime.HasStatusCode(resp, http.StatusOK) {
 		return SQLPoolsClientRenameResponse{}, runtime.NewResponseError(resp)
 	}
-	return SQLPoolsClientRenameResponse{RawResponse: resp}, nil
+	return SQLPoolsClientRenameResponse{}, nil
 }
 
 // renameCreateRequest creates the Rename request.
@@ -433,22 +442,18 @@ func (client *SQLPoolsClient) renameCreateRequest(ctx context.Context, resourceG
 // workspaceName - The name of the workspace.
 // sqlPoolName - SQL pool name
 // options - SQLPoolsClientBeginResumeOptions contains the optional parameters for the SQLPoolsClient.BeginResume method.
-func (client *SQLPoolsClient) BeginResume(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolsClientBeginResumeOptions) (SQLPoolsClientResumePollerResponse, error) {
-	resp, err := client.resume(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
-	if err != nil {
-		return SQLPoolsClientResumePollerResponse{}, err
+func (client *SQLPoolsClient) BeginResume(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolsClientBeginResumeOptions) (*armruntime.Poller[SQLPoolsClientResumeResponse], error) {
+	if options == nil || options.ResumeToken == "" {
+		resp, err := client.resume(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
+		if err != nil {
+			return nil, err
+		}
+		return armruntime.NewPoller(resp, client.pl, &armruntime.NewPollerOptions[SQLPoolsClientResumeResponse]{
+			FinalStateVia: armruntime.FinalStateViaLocation,
+		})
+	} else {
+		return armruntime.NewPollerFromResumeToken[SQLPoolsClientResumeResponse](options.ResumeToken, client.pl, nil)
 	}
-	result := SQLPoolsClientResumePollerResponse{
-		RawResponse: resp,
-	}
-	pt, err := armruntime.NewPoller("SQLPoolsClient.Resume", "location", resp, client.pl)
-	if err != nil {
-		return SQLPoolsClientResumePollerResponse{}, err
-	}
-	result.Poller = &SQLPoolsClientResumePoller{
-		pt: pt,
-	}
-	return result, nil
 }
 
 // Resume - Resume a SQL pool
@@ -552,7 +557,7 @@ func (client *SQLPoolsClient) updateCreateRequest(ctx context.Context, resourceG
 
 // updateHandleResponse handles the Update response.
 func (client *SQLPoolsClient) updateHandleResponse(resp *http.Response) (SQLPoolsClientUpdateResponse, error) {
-	result := SQLPoolsClientUpdateResponse{RawResponse: resp}
+	result := SQLPoolsClientUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPool); err != nil {
 		return SQLPoolsClientUpdateResponse{}, err
 	}
