@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type SQLPoolTableColumnsClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLPoolTableColumnsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SQLPoolTableColumnsClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewSQLPoolTableColumnsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLPoolTableColumnsClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLPoolTableColumnsClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // ListByTableName - Gets columns in a given table in a SQL pool.
@@ -58,16 +63,32 @@ func NewSQLPoolTableColumnsClient(subscriptionID string, credential azcore.Token
 // tableName - The name of the table.
 // options - SQLPoolTableColumnsClientListByTableNameOptions contains the optional parameters for the SQLPoolTableColumnsClient.ListByTableName
 // method.
-func (client *SQLPoolTableColumnsClient) ListByTableName(resourceGroupName string, workspaceName string, sqlPoolName string, schemaName string, tableName string, options *SQLPoolTableColumnsClientListByTableNameOptions) *SQLPoolTableColumnsClientListByTableNamePager {
-	return &SQLPoolTableColumnsClientListByTableNamePager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listByTableNameCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, schemaName, tableName, options)
+func (client *SQLPoolTableColumnsClient) ListByTableName(resourceGroupName string, workspaceName string, sqlPoolName string, schemaName string, tableName string, options *SQLPoolTableColumnsClientListByTableNameOptions) *runtime.Pager[SQLPoolTableColumnsClientListByTableNameResponse] {
+	return runtime.NewPager(runtime.PageProcessor[SQLPoolTableColumnsClientListByTableNameResponse]{
+		More: func(page SQLPoolTableColumnsClientListByTableNameResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp SQLPoolTableColumnsClientListByTableNameResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.SQLPoolColumnListResult.NextLink)
+		Fetcher: func(ctx context.Context, page *SQLPoolTableColumnsClientListByTableNameResponse) (SQLPoolTableColumnsClientListByTableNameResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listByTableNameCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, schemaName, tableName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return SQLPoolTableColumnsClientListByTableNameResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return SQLPoolTableColumnsClientListByTableNameResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return SQLPoolTableColumnsClientListByTableNameResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listByTableNameHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listByTableNameCreateRequest creates the ListByTableName request.
@@ -113,7 +134,7 @@ func (client *SQLPoolTableColumnsClient) listByTableNameCreateRequest(ctx contex
 
 // listByTableNameHandleResponse handles the ListByTableName response.
 func (client *SQLPoolTableColumnsClient) listByTableNameHandleResponse(resp *http.Response) (SQLPoolTableColumnsClientListByTableNameResponse, error) {
-	result := SQLPoolTableColumnsClientListByTableNameResponse{RawResponse: resp}
+	result := SQLPoolTableColumnsClientListByTableNameResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPoolColumnListResult); err != nil {
 		return SQLPoolTableColumnsClientListByTableNameResponse{}, err
 	}
