@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type SQLPoolTransparentDataEncryptionsClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLPoolTransparentDataEncryptionsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SQLPoolTransparentDataEncryptionsClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewSQLPoolTransparentDataEncryptionsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLPoolTransparentDataEncryptionsClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLPoolTransparentDataEncryptionsClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CreateOrUpdate - Creates or updates a Sql pool's transparent data encryption configuration.
@@ -109,7 +114,7 @@ func (client *SQLPoolTransparentDataEncryptionsClient) createOrUpdateCreateReque
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
 func (client *SQLPoolTransparentDataEncryptionsClient) createOrUpdateHandleResponse(resp *http.Response) (SQLPoolTransparentDataEncryptionsClientCreateOrUpdateResponse, error) {
-	result := SQLPoolTransparentDataEncryptionsClientCreateOrUpdateResponse{RawResponse: resp}
+	result := SQLPoolTransparentDataEncryptionsClientCreateOrUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.TransparentDataEncryption); err != nil {
 		return SQLPoolTransparentDataEncryptionsClientCreateOrUpdateResponse{}, err
 	}
@@ -175,7 +180,7 @@ func (client *SQLPoolTransparentDataEncryptionsClient) getCreateRequest(ctx cont
 
 // getHandleResponse handles the Get response.
 func (client *SQLPoolTransparentDataEncryptionsClient) getHandleResponse(resp *http.Response) (SQLPoolTransparentDataEncryptionsClientGetResponse, error) {
-	result := SQLPoolTransparentDataEncryptionsClientGetResponse{RawResponse: resp}
+	result := SQLPoolTransparentDataEncryptionsClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.TransparentDataEncryption); err != nil {
 		return SQLPoolTransparentDataEncryptionsClientGetResponse{}, err
 	}
@@ -189,16 +194,32 @@ func (client *SQLPoolTransparentDataEncryptionsClient) getHandleResponse(resp *h
 // sqlPoolName - SQL pool name
 // options - SQLPoolTransparentDataEncryptionsClientListOptions contains the optional parameters for the SQLPoolTransparentDataEncryptionsClient.List
 // method.
-func (client *SQLPoolTransparentDataEncryptionsClient) List(resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolTransparentDataEncryptionsClientListOptions) *SQLPoolTransparentDataEncryptionsClientListPager {
-	return &SQLPoolTransparentDataEncryptionsClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
+func (client *SQLPoolTransparentDataEncryptionsClient) List(resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolTransparentDataEncryptionsClientListOptions) *runtime.Pager[SQLPoolTransparentDataEncryptionsClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[SQLPoolTransparentDataEncryptionsClientListResponse]{
+		More: func(page SQLPoolTransparentDataEncryptionsClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp SQLPoolTransparentDataEncryptionsClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.TransparentDataEncryptionListResult.NextLink)
+		Fetcher: func(ctx context.Context, page *SQLPoolTransparentDataEncryptionsClientListResponse) (SQLPoolTransparentDataEncryptionsClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return SQLPoolTransparentDataEncryptionsClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return SQLPoolTransparentDataEncryptionsClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return SQLPoolTransparentDataEncryptionsClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -233,7 +254,7 @@ func (client *SQLPoolTransparentDataEncryptionsClient) listCreateRequest(ctx con
 
 // listHandleResponse handles the List response.
 func (client *SQLPoolTransparentDataEncryptionsClient) listHandleResponse(resp *http.Response) (SQLPoolTransparentDataEncryptionsClientListResponse, error) {
-	result := SQLPoolTransparentDataEncryptionsClientListResponse{RawResponse: resp}
+	result := SQLPoolTransparentDataEncryptionsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.TransparentDataEncryptionListResult); err != nil {
 		return SQLPoolTransparentDataEncryptionsClientListResponse{}, err
 	}
