@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type SQLPoolVulnerabilityAssessmentRuleBaselinesClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLPoolVulnerabilityAssessmentRuleBaselinesClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SQLPoolVulnerabilityAssessmentRuleBaselinesClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewSQLPoolVulnerabilityAssessmentRuleBaselinesClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLPoolVulnerabilityAssessmentRuleBaselinesClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLPoolVulnerabilityAssessmentRuleBaselinesClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CreateOrUpdate - Creates or updates a Sql pool's vulnerability assessment rule baseline.
@@ -120,7 +125,7 @@ func (client *SQLPoolVulnerabilityAssessmentRuleBaselinesClient) createOrUpdateC
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
 func (client *SQLPoolVulnerabilityAssessmentRuleBaselinesClient) createOrUpdateHandleResponse(resp *http.Response) (SQLPoolVulnerabilityAssessmentRuleBaselinesClientCreateOrUpdateResponse, error) {
-	result := SQLPoolVulnerabilityAssessmentRuleBaselinesClientCreateOrUpdateResponse{RawResponse: resp}
+	result := SQLPoolVulnerabilityAssessmentRuleBaselinesClientCreateOrUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPoolVulnerabilityAssessmentRuleBaseline); err != nil {
 		return SQLPoolVulnerabilityAssessmentRuleBaselinesClientCreateOrUpdateResponse{}, err
 	}
@@ -150,7 +155,7 @@ func (client *SQLPoolVulnerabilityAssessmentRuleBaselinesClient) Delete(ctx cont
 	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusNoContent) {
 		return SQLPoolVulnerabilityAssessmentRuleBaselinesClientDeleteResponse{}, runtime.NewResponseError(resp)
 	}
-	return SQLPoolVulnerabilityAssessmentRuleBaselinesClientDeleteResponse{RawResponse: resp}, nil
+	return SQLPoolVulnerabilityAssessmentRuleBaselinesClientDeleteResponse{}, nil
 }
 
 // deleteCreateRequest creates the Delete request.
@@ -264,7 +269,7 @@ func (client *SQLPoolVulnerabilityAssessmentRuleBaselinesClient) getCreateReques
 
 // getHandleResponse handles the Get response.
 func (client *SQLPoolVulnerabilityAssessmentRuleBaselinesClient) getHandleResponse(resp *http.Response) (SQLPoolVulnerabilityAssessmentRuleBaselinesClientGetResponse, error) {
-	result := SQLPoolVulnerabilityAssessmentRuleBaselinesClientGetResponse{RawResponse: resp}
+	result := SQLPoolVulnerabilityAssessmentRuleBaselinesClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPoolVulnerabilityAssessmentRuleBaseline); err != nil {
 		return SQLPoolVulnerabilityAssessmentRuleBaselinesClientGetResponse{}, err
 	}
