@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type SQLPoolVulnerabilityAssessmentsClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLPoolVulnerabilityAssessmentsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SQLPoolVulnerabilityAssessmentsClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewSQLPoolVulnerabilityAssessmentsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLPoolVulnerabilityAssessmentsClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLPoolVulnerabilityAssessmentsClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CreateOrUpdate - Creates or updates the Sql pool vulnerability assessment.
@@ -109,7 +114,7 @@ func (client *SQLPoolVulnerabilityAssessmentsClient) createOrUpdateCreateRequest
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
 func (client *SQLPoolVulnerabilityAssessmentsClient) createOrUpdateHandleResponse(resp *http.Response) (SQLPoolVulnerabilityAssessmentsClientCreateOrUpdateResponse, error) {
-	result := SQLPoolVulnerabilityAssessmentsClientCreateOrUpdateResponse{RawResponse: resp}
+	result := SQLPoolVulnerabilityAssessmentsClientCreateOrUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPoolVulnerabilityAssessment); err != nil {
 		return SQLPoolVulnerabilityAssessmentsClientCreateOrUpdateResponse{}, err
 	}
@@ -136,7 +141,7 @@ func (client *SQLPoolVulnerabilityAssessmentsClient) Delete(ctx context.Context,
 	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusNoContent) {
 		return SQLPoolVulnerabilityAssessmentsClientDeleteResponse{}, runtime.NewResponseError(resp)
 	}
-	return SQLPoolVulnerabilityAssessmentsClientDeleteResponse{RawResponse: resp}, nil
+	return SQLPoolVulnerabilityAssessmentsClientDeleteResponse{}, nil
 }
 
 // deleteCreateRequest creates the Delete request.
@@ -231,7 +236,7 @@ func (client *SQLPoolVulnerabilityAssessmentsClient) getCreateRequest(ctx contex
 
 // getHandleResponse handles the Get response.
 func (client *SQLPoolVulnerabilityAssessmentsClient) getHandleResponse(resp *http.Response) (SQLPoolVulnerabilityAssessmentsClientGetResponse, error) {
-	result := SQLPoolVulnerabilityAssessmentsClientGetResponse{RawResponse: resp}
+	result := SQLPoolVulnerabilityAssessmentsClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPoolVulnerabilityAssessment); err != nil {
 		return SQLPoolVulnerabilityAssessmentsClientGetResponse{}, err
 	}
@@ -245,16 +250,32 @@ func (client *SQLPoolVulnerabilityAssessmentsClient) getHandleResponse(resp *htt
 // sqlPoolName - SQL pool name
 // options - SQLPoolVulnerabilityAssessmentsClientListOptions contains the optional parameters for the SQLPoolVulnerabilityAssessmentsClient.List
 // method.
-func (client *SQLPoolVulnerabilityAssessmentsClient) List(resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolVulnerabilityAssessmentsClientListOptions) *SQLPoolVulnerabilityAssessmentsClientListPager {
-	return &SQLPoolVulnerabilityAssessmentsClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
+func (client *SQLPoolVulnerabilityAssessmentsClient) List(resourceGroupName string, workspaceName string, sqlPoolName string, options *SQLPoolVulnerabilityAssessmentsClientListOptions) *runtime.Pager[SQLPoolVulnerabilityAssessmentsClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[SQLPoolVulnerabilityAssessmentsClientListResponse]{
+		More: func(page SQLPoolVulnerabilityAssessmentsClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp SQLPoolVulnerabilityAssessmentsClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.SQLPoolVulnerabilityAssessmentListResult.NextLink)
+		Fetcher: func(ctx context.Context, page *SQLPoolVulnerabilityAssessmentsClientListResponse) (SQLPoolVulnerabilityAssessmentsClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return SQLPoolVulnerabilityAssessmentsClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return SQLPoolVulnerabilityAssessmentsClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return SQLPoolVulnerabilityAssessmentsClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -289,7 +310,7 @@ func (client *SQLPoolVulnerabilityAssessmentsClient) listCreateRequest(ctx conte
 
 // listHandleResponse handles the List response.
 func (client *SQLPoolVulnerabilityAssessmentsClient) listHandleResponse(resp *http.Response) (SQLPoolVulnerabilityAssessmentsClientListResponse, error) {
-	result := SQLPoolVulnerabilityAssessmentsClientListResponse{RawResponse: resp}
+	result := SQLPoolVulnerabilityAssessmentsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPoolVulnerabilityAssessmentListResult); err != nil {
 		return SQLPoolVulnerabilityAssessmentsClientListResponse{}, err
 	}
