@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type SQLPoolVulnerabilityAssessmentScansClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewSQLPoolVulnerabilityAssessmentScansClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SQLPoolVulnerabilityAssessmentScansClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewSQLPoolVulnerabilityAssessmentScansClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*SQLPoolVulnerabilityAssessmentScansClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &SQLPoolVulnerabilityAssessmentScansClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // Export - Convert an existing scan result to a human readable format. If already exists nothing happens
@@ -113,7 +118,7 @@ func (client *SQLPoolVulnerabilityAssessmentScansClient) exportCreateRequest(ctx
 
 // exportHandleResponse handles the Export response.
 func (client *SQLPoolVulnerabilityAssessmentScansClient) exportHandleResponse(resp *http.Response) (SQLPoolVulnerabilityAssessmentScansClientExportResponse, error) {
-	result := SQLPoolVulnerabilityAssessmentScansClientExportResponse{RawResponse: resp}
+	result := SQLPoolVulnerabilityAssessmentScansClientExportResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SQLPoolVulnerabilityAssessmentScansExport); err != nil {
 		return SQLPoolVulnerabilityAssessmentScansClientExportResponse{}, err
 	}
@@ -184,7 +189,7 @@ func (client *SQLPoolVulnerabilityAssessmentScansClient) getCreateRequest(ctx co
 
 // getHandleResponse handles the Get response.
 func (client *SQLPoolVulnerabilityAssessmentScansClient) getHandleResponse(resp *http.Response) (SQLPoolVulnerabilityAssessmentScansClientGetResponse, error) {
-	result := SQLPoolVulnerabilityAssessmentScansClientGetResponse{RawResponse: resp}
+	result := SQLPoolVulnerabilityAssessmentScansClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.VulnerabilityAssessmentScanRecord); err != nil {
 		return SQLPoolVulnerabilityAssessmentScansClientGetResponse{}, err
 	}
@@ -200,22 +205,16 @@ func (client *SQLPoolVulnerabilityAssessmentScansClient) getHandleResponse(resp 
 // scanID - The vulnerability assessment scan Id of the scan to retrieve.
 // options - SQLPoolVulnerabilityAssessmentScansClientBeginInitiateScanOptions contains the optional parameters for the SQLPoolVulnerabilityAssessmentScansClient.BeginInitiateScan
 // method.
-func (client *SQLPoolVulnerabilityAssessmentScansClient) BeginInitiateScan(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, vulnerabilityAssessmentName VulnerabilityAssessmentName, scanID string, options *SQLPoolVulnerabilityAssessmentScansClientBeginInitiateScanOptions) (SQLPoolVulnerabilityAssessmentScansClientInitiateScanPollerResponse, error) {
-	resp, err := client.initiateScan(ctx, resourceGroupName, workspaceName, sqlPoolName, vulnerabilityAssessmentName, scanID, options)
-	if err != nil {
-		return SQLPoolVulnerabilityAssessmentScansClientInitiateScanPollerResponse{}, err
+func (client *SQLPoolVulnerabilityAssessmentScansClient) BeginInitiateScan(ctx context.Context, resourceGroupName string, workspaceName string, sqlPoolName string, vulnerabilityAssessmentName VulnerabilityAssessmentName, scanID string, options *SQLPoolVulnerabilityAssessmentScansClientBeginInitiateScanOptions) (*armruntime.Poller[SQLPoolVulnerabilityAssessmentScansClientInitiateScanResponse], error) {
+	if options == nil || options.ResumeToken == "" {
+		resp, err := client.initiateScan(ctx, resourceGroupName, workspaceName, sqlPoolName, vulnerabilityAssessmentName, scanID, options)
+		if err != nil {
+			return nil, err
+		}
+		return armruntime.NewPoller[SQLPoolVulnerabilityAssessmentScansClientInitiateScanResponse](resp, client.pl, nil)
+	} else {
+		return armruntime.NewPollerFromResumeToken[SQLPoolVulnerabilityAssessmentScansClientInitiateScanResponse](options.ResumeToken, client.pl, nil)
 	}
-	result := SQLPoolVulnerabilityAssessmentScansClientInitiateScanPollerResponse{
-		RawResponse: resp,
-	}
-	pt, err := armruntime.NewPoller("SQLPoolVulnerabilityAssessmentScansClient.InitiateScan", "", resp, client.pl)
-	if err != nil {
-		return SQLPoolVulnerabilityAssessmentScansClientInitiateScanPollerResponse{}, err
-	}
-	result.Poller = &SQLPoolVulnerabilityAssessmentScansClientInitiateScanPoller{
-		pt: pt,
-	}
-	return result, nil
 }
 
 // InitiateScan - Executes a Vulnerability Assessment database scan.
@@ -280,16 +279,32 @@ func (client *SQLPoolVulnerabilityAssessmentScansClient) initiateScanCreateReque
 // vulnerabilityAssessmentName - The name of the vulnerability assessment.
 // options - SQLPoolVulnerabilityAssessmentScansClientListOptions contains the optional parameters for the SQLPoolVulnerabilityAssessmentScansClient.List
 // method.
-func (client *SQLPoolVulnerabilityAssessmentScansClient) List(resourceGroupName string, workspaceName string, sqlPoolName string, vulnerabilityAssessmentName VulnerabilityAssessmentName, options *SQLPoolVulnerabilityAssessmentScansClientListOptions) *SQLPoolVulnerabilityAssessmentScansClientListPager {
-	return &SQLPoolVulnerabilityAssessmentScansClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, vulnerabilityAssessmentName, options)
+func (client *SQLPoolVulnerabilityAssessmentScansClient) List(resourceGroupName string, workspaceName string, sqlPoolName string, vulnerabilityAssessmentName VulnerabilityAssessmentName, options *SQLPoolVulnerabilityAssessmentScansClientListOptions) *runtime.Pager[SQLPoolVulnerabilityAssessmentScansClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[SQLPoolVulnerabilityAssessmentScansClientListResponse]{
+		More: func(page SQLPoolVulnerabilityAssessmentScansClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp SQLPoolVulnerabilityAssessmentScansClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.VulnerabilityAssessmentScanRecordListResult.NextLink)
+		Fetcher: func(ctx context.Context, page *SQLPoolVulnerabilityAssessmentScansClientListResponse) (SQLPoolVulnerabilityAssessmentScansClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, resourceGroupName, workspaceName, sqlPoolName, vulnerabilityAssessmentName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return SQLPoolVulnerabilityAssessmentScansClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return SQLPoolVulnerabilityAssessmentScansClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return SQLPoolVulnerabilityAssessmentScansClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -328,7 +343,7 @@ func (client *SQLPoolVulnerabilityAssessmentScansClient) listCreateRequest(ctx c
 
 // listHandleResponse handles the List response.
 func (client *SQLPoolVulnerabilityAssessmentScansClient) listHandleResponse(resp *http.Response) (SQLPoolVulnerabilityAssessmentScansClientListResponse, error) {
-	result := SQLPoolVulnerabilityAssessmentScansClientListResponse{RawResponse: resp}
+	result := SQLPoolVulnerabilityAssessmentScansClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.VulnerabilityAssessmentScanRecordListResult); err != nil {
 		return SQLPoolVulnerabilityAssessmentScansClientListResponse{}, err
 	}
