@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -33,20 +34,24 @@ type WorkspaceManagedSQLServerVulnerabilityAssessmentsClient struct {
 // subscriptionID - The ID of the target subscription.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewWorkspaceManagedSQLServerVulnerabilityAssessmentsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewWorkspaceManagedSQLServerVulnerabilityAssessmentsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*WorkspaceManagedSQLServerVulnerabilityAssessmentsClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &WorkspaceManagedSQLServerVulnerabilityAssessmentsClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CreateOrUpdate - Create or Update workspace managed sql server's vulnerability assessment.
@@ -104,7 +109,7 @@ func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) createOrU
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
 func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) createOrUpdateHandleResponse(resp *http.Response) (WorkspaceManagedSQLServerVulnerabilityAssessmentsClientCreateOrUpdateResponse, error) {
-	result := WorkspaceManagedSQLServerVulnerabilityAssessmentsClientCreateOrUpdateResponse{RawResponse: resp}
+	result := WorkspaceManagedSQLServerVulnerabilityAssessmentsClientCreateOrUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ServerVulnerabilityAssessment); err != nil {
 		return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientCreateOrUpdateResponse{}, err
 	}
@@ -130,7 +135,7 @@ func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) Delete(ct
 	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusNoContent) {
 		return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientDeleteResponse{}, runtime.NewResponseError(resp)
 	}
-	return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientDeleteResponse{RawResponse: resp}, nil
+	return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientDeleteResponse{}, nil
 }
 
 // deleteCreateRequest creates the Delete request.
@@ -216,7 +221,7 @@ func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) getCreate
 
 // getHandleResponse handles the Get response.
 func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) getHandleResponse(resp *http.Response) (WorkspaceManagedSQLServerVulnerabilityAssessmentsClientGetResponse, error) {
-	result := WorkspaceManagedSQLServerVulnerabilityAssessmentsClientGetResponse{RawResponse: resp}
+	result := WorkspaceManagedSQLServerVulnerabilityAssessmentsClientGetResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ServerVulnerabilityAssessment); err != nil {
 		return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientGetResponse{}, err
 	}
@@ -229,16 +234,32 @@ func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) getHandle
 // workspaceName - The name of the workspace.
 // options - WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListOptions contains the optional parameters for the WorkspaceManagedSQLServerVulnerabilityAssessmentsClient.List
 // method.
-func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) List(resourceGroupName string, workspaceName string, options *WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListOptions) *WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListPager {
-	return &WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, resourceGroupName, workspaceName, options)
+func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) List(resourceGroupName string, workspaceName string, options *WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListOptions) *runtime.Pager[WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse]{
+		More: func(page WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.ServerVulnerabilityAssessmentListResult.NextLink)
+		Fetcher: func(ctx context.Context, page *WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse) (WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, resourceGroupName, workspaceName, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -269,7 +290,7 @@ func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) listCreat
 
 // listHandleResponse handles the List response.
 func (client *WorkspaceManagedSQLServerVulnerabilityAssessmentsClient) listHandleResponse(resp *http.Response) (WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse, error) {
-	result := WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse{RawResponse: resp}
+	result := WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ServerVulnerabilityAssessmentListResult); err != nil {
 		return WorkspaceManagedSQLServerVulnerabilityAssessmentsClientListResponse{}, err
 	}
